[Add] New Process Class
This is an implementation of another design of the Process Class. The idea is to stay closer to the general paper and formulation of what a process is. As far as I understood a Process is the logical combination of all rules that are in the Declare KB. Hence, there is alway only one Process with a bunch of automata that represent the rules from the Declare KB.